gemmbench/requirements.txt
attrs==22.2.0
backcall==0.2.0
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.38.0
h5py
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-client==7.4.8
jupyter-core==4.12.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
nbformat==5.7.2
nest-asyncio==1.5.6
numpy==1.21.6
packaging==23.0
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
plotly==5.11.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
six==1.16.0
tenacity==8.1.0
tornado==6.2
tqdm
traitlets==5.8.1
typing-extensions==4.4.0
wcwidth==0.2.5
widgetsnbextension==4.0.5
zipp==3.11.0